use shared upload service for predictor parsing
