+++ src/app/upload-data/upload-data-v2.service.ts
@@ -18,7 +18,6 @@ import { UtilityMeterDatadbService } from '../indexedDB/utilityMeterData-db.serv
 import { getMeterDataCopy } from '../calculations/conversions/convertMeterData';
 import { ConvertValue } from '../calculations/conversions/convertValue';
 import { GlobalWarmingPotential, GlobalWarmingPotentials } from '../models/globalWarmingPotentials';
-import { PredictordbService } from '../indexedDB/predictors-db.service';
 
 @Injectable({
   providedIn: 'root'
@@ -31,8 +30,7 @@ export class UploadDataV2Service {
     private utilityMeterDbService: UtilityMeterdbService,
     private uploadDataSharedFunctionsService: UploadDataSharedFunctionsService,
     private editMeterFormService: EditMeterFormService,
-    private utilityMeterDataDbService: UtilityMeterDatadbService,
-    private predictorDbService: PredictordbService) { }
+    private utilityMeterDataDbService: UtilityMeterDatadbService) { }
 
 
   parseTemplate(workbook: XLSX.WorkBook): ParsedTemplate {
@@ -42,7 +40,7 @@ export class UploadDataV2Service {
       throw('No Facilities Found!')
     }else{
       let importMetersAndGroups: { meters: Array<IdbUtilityMeter>, newGroups: Array<IdbUtilityMeterGroup> } = this.getImportMeters(workbook, importFacilities, selectedAccount);
-      let predictorEntries: Array<IdbPredictorEntry> = this.getPredictorData(workbook, importFacilities, selectedAccount);
+      let predictorEntries: Array<IdbPredictorEntry> = this.uploadDataSharedFunctionsService.getPredictorData(workbook, importFacilities, selectedAccount);
       let importMeterData: Array<IdbUtilityMeterData> = this.getUtilityMeterData(workbook, importMetersAndGroups.meters);  
       return { importFacilities: importFacilities, importMeters: importMetersAndGroups.meters, predictorEntries: predictorEntries, meterData: importMeterData, newGroups: importMetersAndGroups.newGroups }
     }
@@ -518,72 +516,5 @@ export class UploadDataV2Service {
       globalWarmingPotential: undefined,
       globalWarmingPotentialOption: undefined
     }
-  }
-
-  getPredictorData(workbook: XLSX.WorkBook, importFacilities: Array<IdbFacility>, selectedAccount: IdbAccount): Array<IdbPredictorEntry> {
-    let predictorsData = XLSX.utils.sheet_to_json(workbook.Sheets['Predictors']);
-    let predictorEntries: Array<IdbPredictorEntry> = new Array();
-    let accountPredictorEntries: Array<IdbPredictorEntry> = this.predictorDbService.getAccountPerdictorsCopy();
-    importFacilities.forEach(facility => {
-      let facilityPredictorData = predictorsData.filter(data => { return data['Facility Name'] == facility.name });
-      let facilityPredictorEntries: Array<IdbPredictorEntry> = accountPredictorEntries.filter(entry => { return entry.facilityId == facility.guid });
-      let existingFacilityPredictorData: Array<PredictorData> = new Array();
-      if (facilityPredictorEntries.length != 0) {
-        existingFacilityPredictorData = facilityPredictorEntries[0].predictors.map(predictor => { return predictor });
-      }
-      if (facilityPredictorData.length != 0) {
-        Object.keys(facilityPredictorData[0]).forEach((key) => {
-          if (key != 'Facility Name' && key != 'Date') {
-            let predictorIndex: number = existingFacilityPredictorData.findIndex(predictor => { return predictor.name == key });
-            if (predictorIndex == -1) {
-              let hasData: boolean = false;
-              facilityPredictorData.forEach(dataItem => {
-                if (dataItem[key] != 0) {
-                  hasData = true;
-                }
-              });
-              if (hasData) {
-                let newPredictor: PredictorData = this.predictorDbService.getNewPredictor([]);
-                let nameTest: string = key.toLocaleLowerCase();
-                if (!nameTest.includes('cdd') && !nameTest.includes('hdd')) {
-                  newPredictor.production = true;
-                }
-                newPredictor.name = key;
-                existingFacilityPredictorData.push(newPredictor);
-                facilityPredictorEntries.forEach(predictorEntry => {
-                  predictorEntry.predictors.push(newPredictor);
-                });
-              }
-            }
-          }
-        });
-      }
-      facilityPredictorData.forEach(dataItem => {
-        let dataItemDate: Date = new Date(dataItem['Date']);
-        let facilityPredictorEntry: IdbPredictorEntry = facilityPredictorEntries.find(entry => {
-          return checkSameMonth(dataItemDate, new Date(entry.date))
-        });
-        if (!facilityPredictorEntry) {
-          facilityPredictorEntry = this.predictorDbService.getNewIdbPredictorEntry(facility.guid, selectedAccount.guid, dataItemDate);
-          if (facilityPredictorEntries.length != 0) {
-            facilityPredictorEntry.predictors = JSON.parse(JSON.stringify(facilityPredictorEntries[0].predictors));
-          } else {
-            facilityPredictorEntry.predictors = JSON.parse(JSON.stringify(existingFacilityPredictorData));
-          }
-        }
-        Object.keys(dataItem).forEach((key) => {
-          if (key != 'Facility Name' && key != 'Date') {
-            let predictorIndex: number = facilityPredictorEntry.predictors.findIndex(predictor => { return predictor.name == key });
-            if (predictorIndex != -1) {
-              facilityPredictorEntry.predictors[predictorIndex].amount = dataItem[key];
-            }
-          }
-        });
-        if (facilityPredictorEntry.predictors.length != 0) {
-          predictorEntries.push(JSON.parse(JSON.stringify(facilityPredictorEntry)));
-        }
-      });
-    })
-    return predictorEntries;
   }
 }
